requirements.txt
streamlit
pandas
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pytest
pytest-mock
pytest-cov
pylint